wip: experimental dtos
