+++ packages/frontend/src/pages/Experimental/types.ts
@@ -1,0 +1,7 @@
+export type VizConfiguration = {
+    libType?: string;
+    vizType?: string;
+    xField?: string;
+    yFields?: string[];
+    pivotFields?: string[];
+};
